Fix crash Serious Editor 64-bit
Fixed level editor crash when calling the function to show all entities <N>.
Now the 64-bit editor has full functionality, just like its 32-bit version.
The cause of the crash was the conversion of pointers to 32 bits.

+++ SamTFE/Sources/WorldEditor/SkyFloat.cpp
@@ -56,7 +56,7 @@ BOOL FloatFromString(HWND hWnd, float &fNumber, BOOL &bValid)
   
   bValid = TRUE;  
   float fTmpNumber = fNumber;
-  int iNumLen, iRetLen;
+  size_t iNumLen, iRetLen;
   iNumLen = strlen( CStringA(szWindowText));
   iRetLen = sscanf( CStringA(szWindowText), "%f", &fTmpNumber);
   if( (iRetLen == 1)  || ((iNumLen == 1) && (szWindowText[0] == '-') || (iNumLen == 0)) )
